fix: unable to click the items in a long list
